LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update.https-expected.txt


PASS update() should succeed when new script is available.
PASS update() should fail when mime type is invalid.
FAIL update() should fail when a response for the main script is redirect. promise_rejects_js: function "function() { throw e }" threw object "SecurityError: Script https://web-platform.test:9443/service-workers/service-worker/resources/update-worker.py?Key=c2e1b806-f9b1-4c31-826f-6cdce89a942d&Mode=redirect load failed" ("SecurityError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS update() should fail when a new script contains a syntax error.
PASS update() should resolve when the install event throws.
PASS update() should fail when the pending uninstall flag is set.
PASS update() should succeed when the script shrinks.
